+++ client/src/api/wrap.js
@@ -1,20 +1,32 @@
 let snackit;
 
-export default function wrap(apicall) {
+export default function wrap(apicall, noError = false) {
   return apicall.then(async (response) => {
     let rep;
     try {
       rep = await response.json();
-    } catch {
-      snackit('Server error');
-      throw new Error('Server error');
+    } catch (err) {
+      if (!noError) {
+        snackit('Server error');
+        throw new Error('Server error');
+      } else {
+        const e = new Error(rep.message);
+        e.status = rep.status;
+        e.code = rep.code;
+        throw e;
+      }
     }
+
     if (response.status == 200) {
       return rep;
     } 
-    snackit(rep.message);
+
+    if (!noError) {
+      snackit(rep.message);
+    }
+    
     const e = new Error(rep.message);
-    e.status = response.status;
+    e.status = rep.status;
     e.code = rep.code;
     throw e;
   });