+++ client/src/components/Logo/Logo.jsx
@@ -2,12 +2,14 @@
 import { useTheme } from '@mui/material/styles';
 import { fontWeight } from '@mui/system';
 
-import logo from '../../assets/images/icons/cosmos.png';
+import logo from '../../assets/images/icons/cosmos_simple_black.png';
+import logoDark from '../../assets/images/icons/cosmos_simple_white.png';
 
 // ==============================|| LOGO SVG ||============================== //
 
 const Logo = () => {
     const theme = useTheme();
+    const isLight = theme.palette.mode === 'light';
 
     return (
         /**
@@ -17,7 +19,7 @@ const Logo = () => {
          *
          */
         <>
-            <img src={logo} alt="Cosmos" width="50" />
+            <img src={isLight ? logo : logoDark} alt="Cosmos" width="40" />
             <span style={{fontWeight: 'bold', fontSize: '170%', paddingLeft:'10px'}}> Cosmos</span>
         </>
     );