+++ client/src/layout/MainLayout/Drawer/DrawerContent/Navigation/NavItem.jsx
@@ -23,7 +23,9 @@ const NavItem = ({ item, level }) => {
         itemTarget = '_blank';
     }
 
-    let listItemProps = { component: forwardRef((props, ref) => <Link ref={ref} {...props} to={item.url} target={itemTarget} />) };
+    let listItemProps = { component: forwardRef((props, ref) => <Link ref={ref} {...props} to={item.url} target={itemTarget}
+        rel={itemTarget === '_blank' ? 'noopener noreferrer' : ''} 
+     />) };
     if (item?.external) {
         listItemProps = { component: 'a', href: item.url, target: itemTarget };
     }